# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "postway API",
    "version": "1.0.0",
    "description": "Production-grade serverless REST API with TypeScript and Python Lambda support",
    "contact": {
      "name": "API Support",
      "email": "support@postway.services"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    },
    {
      "url": "https://api-staging.postway.services",
      "description": "Staging environment"
    },
    {
      "url": "https://api.postway.services",
      "description": "Production environment"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "WorkOS JWT token obtained from authentication"
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string",
                "example": "Error message"
              },
              "code": {
                "type": "string",
                "example": "ERROR_CODE"
              }
            }
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
    },
    "responses": {
...
      "Unauthorized": {
        "description": "Unauthorized - Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "message": "Authentication required",
                "code": "UNAUTHORIZED"
              }
            }
          }
        }
      },
...
      "ServerError": {
        "description": "Internal Server Error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "success": false,
              "error": {
                "message": "Internal server error",
                "code": "INTERNAL_ERROR"
              }
            }
          }
        }
      }
    }
  },
...
  "paths": {
...
    "/v1/webhooks/workos": {
      "post": {
        "tags": [
          "Webhooks"
        ],
        "summary": "WorkOS webhook handler",
        "description": "Handles WorkOS webhook events for user and organization lifecycle management.\nVerifies webhook signature and processes events idempotently.\n\n**Supported Events:**\n- `user.created` - Creates new user and auth identity\n- `user.updated` - Updates existing user data\n- `user.deleted` - Removes user and auth identity\n- `organization.created` - Creates new organization\n- `organization.updated` - Updates organization data\n- `organization.deleted` - Removes organization\n\n**Security:** Requires valid WorkOS webhook signature in headers.\n",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string",
                    "description": "Unique event ID",
                    "example": "evt_01H1234567890ABCDEFGHIJK"
                  },
                  "event": {
                    "type": "string",
                    "description": "Event type",
                    "enum": [
                      "user.created",
                      "user.updated",
                      "user.deleted",
                      "organization.created",
                      "organization.updated",
                      "organization.deleted"
                    ],
                    "example": "user.created"
                  },
                  "data": {
                    "type": "object",
                    "description": "Event payload (varies by event type)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook processed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "enum": [
                            "processed",
                            "already_processed"
                          ]
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/ServerError"
          }
        }
      }
    }
  }
}